pull out checkout form logic into it's own component used in app.js
